Research_AI_law/FASE4/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "fase4-eda-principal"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "pandera>=0.18",
    "pyyaml>=6.0",
    "openpyxl>=3.1",
    "pytest>=7.4",
    "seaborn>=0.13",
    "matplotlib>=3.7",
    "plotly>=5.17",
    "scikit-learn>=1.3",
    "statsmodels>=0.14",
    "pingouin>=0.5.4",
    "tqdm>=4.66",
    "kaleido>=0.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
